requirements.txt
streamlit
pandas
faiss-cpu
sentence-transformers
PyPDF2
transformers
